Extracted face code to it's own class. renamed module to face test to let people see the faces if they want.
