+++ FF8/FF8.csproj
@@ -141,6 +141,7 @@
     <Compile Include="ArchiveWorker.cs" />
     <Compile Include="debug_MCH.cs" />
     <Compile Include="ExtapathyExtended.cs" />
+    <Compile Include="Faces.cs" />
     <Compile Include="Ffcc.cs" />
     <Compile Include="FfccVaribleGroup.cs" />
     <Compile Include="FFmpegBinariesHelper.cs" />
@@ -149,7 +150,6 @@
     <Compile Include="init_debugger_Audio.cs" />
     <Compile Include="init_debugger_battle.cs" />
     <Compile Include="init_debugger_fields.cs" />
-    <Compile Include="module_menu_debug.cs" />
     <Compile Include="Input.cs" />
     <Compile Include="LZSS.cs" />
     <Compile Include="MakiExtended.cs" />
@@ -157,6 +157,7 @@
     <Compile Include="ModuleHandler.cs" />
     <Compile Include="module_battle.cs" />
     <Compile Include="module_battle_Debug.cs" />
+    <Compile Include="Module_face_test.cs" />
     <Compile Include="module_field_debug.cs" />
     <Compile Include="module_main_menu_debug.cs" />
     <Compile Include="module_movie_test.cs" />
