+++ FF8/module_main_menu_debug.cs
@@ -24,7 +24,7 @@ namespace FF8
             { Ditems.Music, new Item{Text="Play/Stop music : {0}"} },
             { Ditems.Sounds, new Item{Text="Play audio.dat: {0}"} },
             { Ditems.World, new Item{Text="Jump to World Map"} },
-            { Ditems.Menu, new Item{Text="Test Menu"} }
+            { Ditems.Faces, new Item{Text="Test Faces"} }
         };
 
         /// <summary>
@@ -72,7 +72,7 @@ namespace FF8
             Music,
             Sounds,
             World,
-            Menu
+            Faces
         }
 
         /// <summary>
@@ -544,9 +544,9 @@ namespace FF8
                     Memory.IsMouseVisible = false;
                     break;
 
-                case Ditems.Menu:
+                case Ditems.Faces:
                     Fade = 0.0f;
-                    Memory.module = Memory.MODULE_MENU_DEBUG;
+                    Memory.module = Memory.MODULE_FACE_TEST;
                     break;
 
                 default:
